LFO Module
When the LFO module receives a maximum joystick_y input, the calculated LFO period becomes too short, resulting in a very fast triangle wave. This causes rapid gain fluctuations, which leads to audible distortion and popping noise in the output audio signal.

I generated the bitstream and tested on FPGA. It works but needs to be improved. Currently I'm working to fix timing violation.